sign up form

+++ main.js
@@ -6,29 +6,29 @@ let signupPassword = document.getElementById("signupPassword");
 let userData = [];
 
 function userSignup() {
-  let alertMessage = document.querySelector(".alert-msg");
-  if (validate()) {
-    let user = {
-      userName: signupName.value,
-      userEmail: signupEmail.value,
-      userPassword: signupPassword.value,
-    };
-    userData.push(user);
-    localStorage.setItem("userData", JSON.stringify(userData));
-    clearForm();
-    redirectTologin();
+  if (isEmpty()) {
+    document.getElementById('exist').innerHTML = '<span class="text-danger m-3">All inputs is required</span>'
   } else {
-    alertMessage.classList.replace("d-none", "d-block");
+    if (isEmailExist()) {
+      document.getElementById('exist').innerHTML = '<span class="text-danger m-3">email already exists</span>'
+    } else {
+      if (validate()) {
+        let user = {
+          userName: signupName.value,
+          userEmail: signupEmail.value,
+          userPassword: signupPassword.value,
+        };
+        userData.push(user);
+        localStorage.setItem("userData", JSON.stringify(userData));
+        redirectTologin();
+      } else {
+        document.getElementById('exist').innerHTML = '<span class="text-danger m-3">enter a valid email address</span>'
+      }
+    }
   }
 }
 
 let signupButton = document.querySelector(".signup-btn");
-
-function clearForm() {
-  signupName.value = "";
-  signupEmail.value = "";
-  signupPassword.value = "";
-}
 
 function redirectTologin() {
   location.href = "/index.html";
@@ -37,6 +37,24 @@ function redirectTologin() {
 function validate() {
   const emailRegex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$/;
   return emailRegex.test(signupEmail.value);
+}
+
+function isEmpty() {
+  if (signupName.value == "" || signupEmail.value == "" || signupPassword.value == "") {
+    return true;
+  } else {
+    return false;
+  }
+}
+
+function isEmailExist() {
+  for (var i = 0; i < userData.length; i++) {
+    if (userData[i].email == signupEmail.value) {
+      return true;
+    } else {
+      return false;
+    }
+  }
 }
 
 // user log in section
@@ -57,27 +75,24 @@ let sessionName = [];
 
 function login() {
   for (let i = 0; i < userData.length; i++) {
-    if (
-      userData[i].userEmail.match(loginEmail.toLowerCase()) &&
-      userData[i].userPassword.match(loginPassword.toLowerCase())
-    ) {
+    let email = userData[i].userEmail;
+    let pass = userData[i].userPassword;
+    console.log(email, pass);
+    if (email === loginEmail && pass === loginPassword) {
       localStorage.setItem("sessionName", JSON.stringify(userData[i].userName));
-      
       redirectToHome();
     } else {
-      document.getElementById("incorrect").innerHTML =
-        '<span class="p-2 text-danger">incorrect email or password</span>';
+      document.getElementById("incorrect").innerText =
+        "incorrect email or password";
+      console.log(i);
     }
   }
   function redirectToHome() {
-    location.href = "/home.html";
+    location.href = "/home/home.html";
   }
 }
-let n = JSON.parse(localStorage.getItem("sessionName"))
-console.log(n);
-// document.querySelector(".session-name").innerHTML = ;
-
 let loginButton = document.querySelector(".login-btn");
 loginButton.addEventListener("click", login);
 
-console.log(localStorage.getItem("sessionName"));
+let sessionUser = JSON.parse(localStorage.getItem("sessionName"));
+// document.querySelector(".session-name").innerHTML += sessionUser;
